+++ client/src/pages/Survey/util/Survey.js
@@ -1,0 +1,22 @@
+const surveyQuestions = [
+    {
+        value: "name",
+        question: "What is your adventurers name?",
+        choices: []
+    },
+    {
+        question: "What is your chosen race?",
+        value: "race",
+        choices: ["Dwarf", "Elf", "Halfling", "Human", "Dragonborn", "Gnome", "Half-Elf", "Half-Orc", "Tiefling"]
+    },
+    {
+        question: "What is your chosen class?",
+        value: "class",
+        choices: ["Barbarian", "Bard", "Cleric", "Druid", "Fighter", "Monk", "Paladin", "Ranger", "Rougue", "Sorcerer", "Warlock", "Wizard"]
+    },
+    {
+        question: "What is your chosen alignment?",
+        value: "alignment",
+        choices: ["Lawful Good", "Nuetral Good", "Chaotic Good", "Lawful Nuetral", "True Nuetral", "Chaotic Nuetral", "Lawful Evil", "Nuetral Evil", "Chaotic Evil"]
+    }
+]
